Making MergeJoin and TableScan operations properly report metrics

+++ structured_derby/src/main/java/com/splicemachine/derby/impl/sql/execute/operations/BroadCastJoinRows.java
@@ -51,12 +51,12 @@ public class BroadCastJoinRows implements IJoinRowsIterator<ExecRow> {
     }
 
     @Override
-    public int getLeftRowsSeen() {
+    public long getLeftRowsSeen() {
         return leftRowsSeen;
     }
 
     @Override
-    public int getRightRowsSeen() {
+    public long getRightRowsSeen() {
         return rightRowsSeen;
     }
